+++ js/main.js
@@ -1,98 +1,4 @@
-const MAX_AMOUNT_MESSAGES = 2;
-const MAX_AMOUNT_AVATARS = 6;
-const MIN_AMOUNT_LIKES = 15;
-const MAX_AMOUNT_LIKES = 200;
-const MAX_AMOUNT_COMMENTS = 30;
-const MAX_AMOUNT_PHOTOS = 25;
-
-const namesArray = [
-  'Ева',
-  'Максим',
-  'Полина',
-  'Алиса',
-  'Екатерина',
-  'Артём',
-  'Валерия',
-  'Захар',
-  'Кира',
-  'Александр',
-];
-
-const messagesArray = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const descriptionsArray = [
-  'Мастер быстрых фоток.',
-  'Пришлось проснуться в 4 утра для этого кадра.',
-  'Не могу поверить, что получилось заснять этот момент!',
-  'Ура! Горизонт ровный вышел!',
-  'Любимое время года.'
-];
-
-const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
-
-const getRandomItemFromArray = (array) => array[getRandomNumber(0, array.length - 1)];
-
-const createArrayOfObjects = (length, objConstructor) => Array.from({length}, objConstructor);
-
-const getId = () => {
-  let i = 1;
-  return function () {
-    return i++;
-  };
-};
-
-const getPhotoId = getId();
-const getCommentId = getId();
-
-const getUrl = (folder, format) => function (title) {
-  return `${folder}/${title}.${format}`;
-};
-
-const getPhotoUrl = getUrl('photos', 'jpg');
-const getAvatarUrl = getUrl('img', 'svg');
-
-
-const getMessage = () => {
-  const amount = getRandomNumber(1, MAX_AMOUNT_MESSAGES);
-  const messages = [];
-
-  while (messages.length < amount) {
-    const newMessage = getRandomItemFromArray(messagesArray);
-
-    if (!messages.includes(newMessage)) {
-      messages.push(newMessage);
-    }
-  }
-
-  return messages.join(' ');
-};
-
-const getCommentObj = () => {
-  const title = `avatar-${getRandomNumber(1, MAX_AMOUNT_AVATARS)}`;
-  return {
-    id: getCommentId(),
-    avatar: getAvatarUrl(title),
-    message: getMessage(),
-    name: getRandomItemFromArray(namesArray)
-  };
-};
-
-const getPhotoObj = () => {
-  const id = getPhotoId();
-  return {
-    id,
-    url: getPhotoUrl(id),
-    description: getRandomItemFromArray(descriptionsArray),
-    likes: getRandomNumber(MIN_AMOUNT_LIKES, MAX_AMOUNT_LIKES),
-    comments: createArrayOfObjects(getRandomNumber(0, MAX_AMOUNT_COMMENTS), getCommentObj)
-  };
-};
+import { createArrayOfObjects } from './util.js';
+import { MAX_AMOUNT_PHOTOS, getPhotoObj} from './data.js';
 
 createArrayOfObjects(MAX_AMOUNT_PHOTOS, getPhotoObj);
